[SQL] FLOAT to DECIMAL(15,2) for money

+++ sql/create_table.sql
@@ -447,7 +447,7 @@ CREATE TABLE IF NOT EXISTS `meatballs`.`shift`
 CREATE TABLE IF NOT EXISTS `meatballs`.`facilityBalance`
 (
      f_id        INTEGER NOT NULL PRIMARY KEY,
-    `balance`    FLOAT NOT NULL DEFAULT 0.0,
+    `balance`    DECIMAL (15,2) NOT NULL DEFAULT 0.0,
 
     FOREIGN KEY (`f_id`) REFERENCES `meatballs`.`facility` (`f_id`)
 );
